split the local collection loading logic
